+++ tests/integration/querier_sqlite/testdata/003_nullable_columns/runner.go
@@ -36,13 +36,13 @@ func main() {
 
 	queries := db.New(conn)
 
-	profileWithValues, err := queries.GetProfile(ctx, int32(1))
+	profileWithValues, err := queries.GetProfile(ctx, int64(1))
 	if err != nil {
 		fmt.Fprintln(os.Stderr, err)
 		os.Exit(1)
 	}
 
-	profileWithNulls, err := queries.GetProfile(ctx, int32(2))
+	profileWithNulls, err := queries.GetProfile(ctx, int64(2))
 	if err != nil {
 		fmt.Fprintln(os.Stderr, err)
 		os.Exit(1)
