+++ tests/integration/querier_sqlite/testdata/005_type_mapping/runner.go
@@ -36,7 +36,7 @@ func main() {
 
 	queries := db.New(conn)
 
-	item, err := queries.GetItem(ctx, int32(1))
+	item, err := queries.GetItem(ctx, int64(1))
 	if err != nil {
 		fmt.Fprintln(os.Stderr, err)
 		os.Exit(1)
