+++ tests/integration/querier_sqlite/testdata/007_inner_join/runner.go
@@ -48,7 +48,7 @@ func main() {
 
 	queries := db.New(conn)
 
-	book, err := queries.GetBookWithAuthor(ctx, int32(1))
+	book, err := queries.GetBookWithAuthor(ctx, int64(1))
 	if err != nil {
 		fmt.Fprintln(os.Stderr, err)
 		os.Exit(1)
